Update: modify zsh directory and search patterns
Using our new format, all zsh files are in the zsh directory, which
will be located within .local/share/chezmoi on a given machine.

Also, the naming convention has been modified to keep concerns
separate but not in separate directories. For example,
python.completion.zsh. We've updated the search patterns to allow for
these prefixes.